requirements.txt
beautifulsoup4==4.12.2
python-telegram-bot==20.7
PyYAML==6.0.1
Requests==2.31.0
selenium==4.4.0
Unidecode==1.3.7
webdriver_manager==4.0.1
